+++ examples/src/main/java/tachyon/examples/Performance.java
@@ -15,13 +15,12 @@
 
 package tachyon.examples;
 
-import java.io.InputStream;
 import java.io.IOException;
+import java.io.InputStream;
 import java.io.OutputStream;
 import java.io.RandomAccessFile;
 import java.nio.ByteBuffer;
 import java.nio.ByteOrder;
-import java.nio.IntBuffer;
 import java.nio.channels.FileChannel.MapMode;
 
 import org.apache.hadoop.conf.Configuration;
@@ -37,8 +36,8 @@ import tachyon.Constants;
 import tachyon.TachyonURI;
 import tachyon.Version;
 import tachyon.client.ReadType;
-import tachyon.client.TachyonFile;
 import tachyon.client.TachyonFS;
+import tachyon.client.TachyonFile;
 import tachyon.client.WriteType;
 import tachyon.client.next.file.FileOutStream;
 import tachyon.conf.TachyonConf;
@@ -69,7 +68,7 @@ public class Performance {
   public static void createFiles() throws IOException {
     final long startTimeMs = CommonUtils.getCurrentMs();
     for (int k = 0; k < sFiles; k ++) {
-      int fileId = sMtc.createFile(new TachyonURI(sFileName + (k + sBaseFileNumber)));
+      long fileId = sMtc.createFile(new TachyonURI(sFileName + (k + sBaseFileNumber)));
       LOG.info(FormatUtils.formatTimeTakenMs(startTimeMs,
           "user_createFiles with fileId " + fileId));
     }
@@ -229,7 +228,7 @@ public class Performance {
 
     public void readPartition() throws IOException {
       if (sDebugMode) {
-        ByteBuffer buf = ByteBuffer.allocate((int) sBlockSizeBytes);
+        ByteBuffer buf = ByteBuffer.allocate(sBlockSizeBytes);
         LOG.info("Verifying the reading data...");
 
         for (int pId = mLeft; pId < mRight; pId ++) {
